Replaced static cache instance with singleton per IoC service collection I just realized that the strange interdependencies when running unit tests sequentially (causing tests to fail in AppVeyor, but not on TravisCI or locally) may actually be caused by the global cache instance.
